# Performance Improvements & Some refactor