--- a/regional_standings/model/data_loader.js
+++ b/regional_standings/model/data_loader.js
@@ -1,7 +1,5 @@
 "use strict";
 
-const fs = require('fs');
-const Region = require('./util/region');
 const Team = require('./team');
 const matchData = require('../data/matchdata.json');
 const {SortOrderEnum} = require("./enums/sort_order");