--- a/regional_standings/model/enums/region.js
+++ b/regional_standings/model/enums/region.js
@@ -1,0 +1,11 @@
+"use strict";
+
+const RegionEnum = Object.freeze({
+    EU: 0,
+    AMERICAS: 1,
+    NA: 1,
+    SA: 1,
+    OTHERS: 2,
+});
+
+module.exports = {RegionEnum}