--- a/regional_standings/model/enums/sort_order.js
+++ b/regional_standings/model/enums/sort_order.js
@@ -1,0 +1,10 @@
+"use strict";
+
+const SortOrderEnum = Object.freeze({
+    ASC: 'asc',
+    DESC: 'desc'
+})
+
+module.exports = {
+    SortOrderEnum
+};