--- a/regional_standings/model/enums/text_pad_direction.js
+++ b/regional_standings/model/enums/text_pad_direction.js
@@ -1,0 +1,7 @@
+const TextPadDirection = Object.freeze({
+    LEFT: -1,
+    RIGHT: 1,
+    CENTER: 0,
+})
+
+module.exports = {TextPadDirection};