## Performance Improvements
### main_fit.js +30%
#### Before
- **6.157s**
![image](https://github.com/ValveSoftware/counter-strike/assets/26742503/fb73080c-0afb-417c-922a-9bae04a190f2)

#### After
- **4.272s**
![image](https://github.com/ValveSoftware/counter-strike/assets/26742503/52fc2c8e-26a2-4432-a213-d88092b680e4)

### main.js +27%
#### Before
- **385ms**
![image](https://github.com/ValveSoftware/counter-strike/assets/26742503/37981be0-ab46-47f2-8a94-ff64e650d567)

#### After
- **282ms**
![image](https://github.com/ValveSoftware/counter-strike/assets/26742503/73c23a7b-4c91-4357-9d9a-58b028a0a989)

My machine:
```
Windows 11
Version 22H2 22621.2861)

CPU: i7 10700K
RAM: 48gb
DISC: 1TB SSD NVMe

NodeJS: v18.16.1
```

## Refactor
- Removed empty lines
- Removed unused imports
- Removed unused variables
- Added enum for text pad direction
- Added enum for regions
- Added enum for sort order
- Change some `let` to `const` when possible